Infra: switch to Angular 18 structure
