--- angular/compose.yaml
@@ -1,27 +1,0 @@
-version: '3.9'
-
-services:
-  web:
-    build:
-      context: .
-      target: dev
-    volumes:
-      - .:/app
-      - /tmp/.X11-unix:/tmp/.X11-unix
-    environment:
-      - USER_ID=${UID:-1000}
-      - GROUP_ID=${GID:-1000}
-      - DISPLAY
-      - CHROME_BIN=/usr/bin/chromium-browser
-    ports:
-      - "9876:9876"
-
-  caddy:
-    image: caddy:2.8-alpine
-    volumes:
-      - ./dist/adventrame/browser:/var/www/html
-      - ./Caddyfile:/etc/caddy/Caddyfile
-    ports:
-      - "8085:80"
-    depends_on:
-      - web
