+++ backup/compose.yaml
@@ -1,0 +1,21 @@
+version: "3.9"
+
+services:
+  node:
+    image: node:lts-alpine3.19
+    restart: always
+    volumes:
+      - ./:/app
+    working_dir: /app
+    user: "1000"
+    command: 
+      - /bin/sh
+      - -c
+      - |
+        npm install
+        npm run dev -- --port 80 --host
+    expose:
+      - 80
+    ports:
+      - "8084:80"
+    tty: true
