+++ backup/index.html
@@ -31,7 +31,7 @@
       </form>
     </div>
 
-    <script type="module" src="/src/main.ts"></script>
+    <script type="module" src="/backup/src/src/main.ts"></script>
 </body>
 
 </html>
