+++ backup/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "trame",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "../node_modules/.bin/vite --port 8000 --host",
+    "build": "../node_modules/.bin/tsc && vite build",
+    "preview": "../node_modules/.bin/vite preview",
+    "format": "../node_modules/.bin/prettier --write \\\"src/**/*.ts\\\""
+  },
+  "devDependencies": {
+    "prettier": "3.2.4",
+    "sass": "^1.71.1",
+    "typescript": "^5.2.2",
+    "vite": "^5.0.8"
+  },
+  "dependencies": {
+    "@popperjs/core": "^2.11.8",
+    "@viz-js/viz": "^3.4.0",
+    "bootstrap": "^5.3.3",
+    "normalize.css": "^8.0.1",
+    "pixi.js": "^7.3.3"
+  }
+}
