+++ backup/src/style.scss
@@ -1,5 +1,5 @@
-@import "normalize.css";
-@import "bootstrap/scss/bootstrap";
+@import "../../node_modules/normalize.css";
+@import "../../node_modules/bootstrap/scss/bootstrap";
 
 *, body {
     background-color: #FBFBFB;
